Add files via upload
Fully-coupled structural-acoustic simulation. A harmonic load is applied to the cube which is the structure and the air (cavity) around the cube is meshed. The outer surface of the cube is meshed with acoustic infinite elements to simulate free-field conditions. The linear displacement dof of the cube is tied to the dof of the cavity mesh.